requirements.txt
fastapi
uvicorn[standard]
openai
httpx
requests
python-dotenv
python-multipart
